[FLINK-5129] move path-related methods from BlobUtils to FileSystemBlobStore and cleanup unused methods
